+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Gitaly changelog
 
+UNRELEASED
+
+- Add gitaly-ruby installation debug log messages
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/710
+
 v0.100.0
 
 - Fix WikiFindPage when the page has invalidly-encoded content