+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Gitaly changelog
 
+v0.101.0
+
+Changed
+- Add gitaly-ruby installation debug log messages
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/710
+Fixed
+- Use round robin load balancing instead of 'pick first' for gitaly-ruby
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/700
+Other
+- Generate changelog when releasing a tag to prevent merge conflicts
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/719
+- Unvendor Repository#create implementation
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/713
+
 v0.100.0
 
 - Fix WikiFindPage when the page has invalidly-encoded content